refactor: 인증 관련 인터페이스를 분리하여 다형성 적용
