+++ backend/src/main/java/mouda/backend/auth/Infrastructure/AppleOauthClient.java
@@ -1,0 +1,8 @@
+package mouda.backend.auth.Infrastructure;
+
+public class AppleOauthClient implements OauthClient {
+	@Override
+	public String getIdToken(String code) {
+		return "";
+	}
+}
