+++ backend/src/main/java/mouda/backend/auth/Infrastructure/OauthClient.java
@@ -1,0 +1,6 @@
+package mouda.backend.auth.Infrastructure;
+
+public interface OauthClient {
+
+	String getIdToken(String code);
+}
