+++ backend/src/main/java/mouda/backend/auth/presentation/controller/AuthController.java
@@ -8,7 +8,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import mouda.backend.aop.logging.ExceptRequestLogging;
-import mouda.backend.auth.business.AuthService;
+import mouda.backend.auth.business.AppleAuthService;
+import mouda.backend.auth.business.KakaoAuthService;
 import mouda.backend.auth.presentation.controller.swagger.AuthSwagger;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
@@ -19,12 +20,13 @@ import mouda.backend.common.response.RestResponse;
 @RequiredArgsConstructor
 public class AuthController implements AuthSwagger {
 
-	private final AuthService authService;
+	private final KakaoAuthService kakaoAuthService;
+	private final AppleAuthService appleAuthService;
 
 	@Override
 	@PostMapping("/kakao/oauth")
 	public ResponseEntity<RestResponse<LoginResponse>> loginKakaoOauth(@RequestBody OauthRequest oauthRequest) {
-		LoginResponse response = authService.oauthLogin(oauthRequest);
+		LoginResponse response = kakaoAuthService.oauthLogin(oauthRequest);
 
 		return ResponseEntity.ok().body(new RestResponse<>(response));
 	}
@@ -33,7 +35,7 @@ public class AuthController implements AuthSwagger {
 	@PostMapping("/login")
 	@ExceptRequestLogging
 	public ResponseEntity<RestResponse<LoginResponse>> loginBasicOauth() {
-		LoginResponse response = authService.basicLogin();
+		LoginResponse response = kakaoAuthService.basicLogin();
 
 		return ResponseEntity.ok().body(new RestResponse<>(response));
 	}
