+++ backend/src/main/java/mouda/backend/common/config/argumentresolver/LoginMemberArgumentResolver.java
@@ -1,7 +1,5 @@
 package mouda.backend.common.config.argumentresolver;
 
-import lombok.RequiredArgsConstructor;
-import mouda.backend.auth.business.AuthService;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -9,29 +7,32 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
+import lombok.RequiredArgsConstructor;
+import mouda.backend.auth.business.KakaoAuthService;
+
 @Component
 @RequiredArgsConstructor
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
 
-    private final AuthService authService;
+	private final KakaoAuthService authService;
 
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(LoginMember.class);
-    }
+	@Override
+	public boolean supportsParameter(MethodParameter parameter) {
+		return parameter.hasParameterAnnotation(LoginMember.class);
+	}
 
-    @Override
-    public Object resolveArgument(
-        MethodParameter parameter, ModelAndViewContainer mavContainer,
-        NativeWebRequest webRequest, WebDataBinderFactory binderFactory
-    ) throws Exception {
-        String authorizationHeader = webRequest.getHeader("Authorization");
+	@Override
+	public Object resolveArgument(
+		MethodParameter parameter, ModelAndViewContainer mavContainer,
+		NativeWebRequest webRequest, WebDataBinderFactory binderFactory
+	) throws Exception {
+		String authorizationHeader = webRequest.getHeader("Authorization");
 
-        String token = extractToken(authorizationHeader);
-        return authService.findMember(token);
-    }
+		String token = extractToken(authorizationHeader);
+		return authService.findMember(token);
+	}
 
-    private String extractToken(String authorizationHeader) {
-        return authorizationHeader.substring(7);
-    }
+	private String extractToken(String authorizationHeader) {
+		return authorizationHeader.substring(7);
+	}
 }
