+++ backend/src/main/java/mouda/backend/common/config/interceptor/AuthenticationCheckInterceptor.java
@@ -8,9 +8,9 @@ import org.springframework.web.servlet.HandlerInterceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import mouda.backend.auth.business.AuthService;
 import mouda.backend.auth.exception.AuthErrorMessage;
 import mouda.backend.auth.exception.AuthException;
-import mouda.backend.auth.business.AuthService;
 
 @Component
 @RequiredArgsConstructor
